+++ src/app/shared/PersonName.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import cx from "classnames"
   import {defaultTo} from "ramda"
   import {Directory, user} from "src/app/engine"
 
@@ -8,7 +9,7 @@
   const following = user.followsSet.derived(s => s.has(pubkey))
 </script>
 
-<div class={$$props.class}>
+<div class={cx("flex items-center gap-2", $$props.class)}>
   <span>{Directory.displayProfile($profile)}</span>
   {#if $following}
     <i class="fa fa-user-check ml-2 text-xs text-accent" />